Update server classes.
