+++ pymodbus/server/__init__.py
@@ -22,11 +22,15 @@ __all__ = [
     "get_simulator_commandline",
 ]
 
-from pymodbus.server.async_io import (
+from pymodbus.server.server import (
     ModbusSerialServer,
     ModbusTcpServer,
     ModbusTlsServer,
     ModbusUdpServer,
+)
+from pymodbus.server.simulator.http_server import ModbusSimulatorServer
+from pymodbus.server.simulator.main import get_commandline as get_simulator_commandline
+from pymodbus.server.startstop import (
     ServerAsyncStop,
     ServerStop,
     StartAsyncSerialServer,
@@ -38,5 +42,3 @@ from pymodbus.server.async_io import (
     StartTlsServer,
     StartUdpServer,
 )
-from pymodbus.server.simulator.http_server import ModbusSimulatorServer
-from pymodbus.server.simulator.main import get_commandline as get_simulator_commandline
